backend/requirements.txt
# Core
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
pydantic>=2.7.0,<3.0.0  # Breaking changes entre v2 e v3
pydantic-settings>=2.1.0,<3.0.0

# Database
supabase>=2.27.0,<3.0.0  # Versão compatível com httpx 0.27.x
pgvector>=0.2.0,<0.3.0
psycopg2-binary>=2.9.0,<3.0.0

# Embeddings e RAG
sentence-transformers>=2.2.0,<3.0.0
langchain>=0.3.27,<0.4.0  # Pode ter breaking changes entre versões menores
langchain-community>=0.3.27,<0.4.0

# LLM e APIs
groq>=0.36.0,<1.0.0
google-genai>=0.2.0,<1.0.0  # Nova biblioteca para Gemini (inclui TTS)
# DEPENDÊNCIA OPCIONAL: Para usar GLM 4 (Zhipu AI), instale manualmente:
# pip install "zai-sdk==0.3.1"
# zai-sdk==0.3.1  # Versão testada e estável (descomente se necessário)

# Utilitários
python-dotenv>=1.0.0

# Ferramentas de Desenvolvimento e Qualidade
pytest>=8.0.0
pytest-asyncio>=0.21.0
ruff>=0.1.0
bandit>=1.7.5
requests>=2.31.0  # Para scripts de teste (test_sistema_completo.py)
httpx>=0.26.0,<0.28.0  # Compatível com supabase 2.x (0.28.0+ tem breaking changes)
loguru>=0.7.0
slowapi>=0.1.0
python-multipart>=0.0.6  # Necessário para UploadFile no FastAPI
pybreaker>=1.0.0,<2.0.0  # Circuit breakers para serviços externos

# Audio (se necessário)
pydub>=0.25.0

# Document Processing (MVP 100% FREE - Fase 2)
PyPDF2>=3.0.0      # Extração de texto de PDFs (texto nativo apenas)
pdfplumber>=0.11.0 # Extração melhorada de PDFs com suporte a tabelas (opcional)
openpyxl>=3.1.0    # Para leitura/escrita de Excel
pandas>=2.0.0      # Para processamento de dados Excel

# Document Processing - DOCX/PPTX (Fase 2)
python-docx>=1.1.0    # Para leitura de arquivos DOCX
python-pptx>=0.6.23   # Para leitura de arquivos PPTX

# TODO FUTURO (quando necessário):
# - OCR para PDFs escaneados: Google Document AI (1k páginas/mês free) ou solução paga
# - OCR básico: pytesseract, pdf2image (requer Tesseract instalado no sistema)

